requirements.txt
llama-cpp-python>=0.2.90
pyautogui>=0.9.54
pillow>=10.4.0
pytesseract>=0.3.10
vosk>=0.3.45
sounddevice>=0.4.6
pyttsx3>=2.90
